working for gocode
