+++ src/goDeclaration.ts
